requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
httpx>=0.27.0
pydantic>=2.7.0
python-multipart>=0.0.12
aiohttp>=3.10.0
Pillow>=10.3.0